Suppress SIGPROF
We don't handle it well -- the exact reason is unclear, but threading
seems to trigger it -- so just squelch it.

Building with -pg is necessary for getting call-graph information with
e.g. the Linux 'perf' profiler.  But -pg also causes the libc startup
routine to set a SIGPROF timer.  That doesn't serve any purpose with a
modern profiler like 'perf'; it was used by 'gprof'.